% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bypass_flows}
\alias{bypass_flows}
\title{Flow through the Sutter and Yolo Bypasses}
\format{
a named list where each component contains a dataframe with 972 rows and 7 variables:
\describe{
\item{date}{CALSIM II date}
\item{sutter1}{D117}
\item{sutter2}{C135}
\item{sutter3}{C136A}
\item{sutter4}{C137}
\item{yolo1}{D160}
\item{yolo2}{C157}
}
}
\source{
\itemize{
\item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
\item \strong{Node Selection:} Mark Tompkins \email{mtompkins@flowwest.com}
\item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
bypass_flows
}
\description{
A named list containing datasets with the flow within the bypasses for
estimating available rearing habitat within the bypasses. Each component of the list
contains this output from a different Calsim run:
\itemize{
\item biop_2008_2009: output from the 2008-2009 Calsim II BiOp run.
\item biop_itp_2018_2019: output from the 2018-2019 Calsim II BiOp run.
\item calsim_run_of_river: output from the Calsim Run of River run.
}
}
\details{
The flow in cubic feet per second through the bypasses are represented using
'FLOW-CHANNEL' and 'FLOW-DELIVERY' nodes from CALSIM II.
The nodes for each watershed are outlined above.
}
\section{Model Usage}{

The habitat within the bypasses are split at the major weirs.

\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}

\keyword{datasets}
