% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cfs_to_cms}
\alias{cfs_to_cms}
\title{Convert flow from cfs to cms}
\usage{
cfs_to_cms(flow_cfs)
}
\arguments{
\item{flow_cfs}{flow in cubic feet per second}
}
\value{
\code{flow_cfs} in cubic meters per second
}
\description{
Convert flow from cfs to cms
}
