% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_cross_channel_closed}
\alias{delta_cross_channel_closed}
\title{Delta Cross Channel Operations}
\format{
a 2 by 12 matrix, row one and two are the count and proportion of days
closed respectively, columns are months
}
\source{
\href{http://www.westcoast.fisheries.noaa.gov/central_valley/water_operations/ocap.html}{2009 NMFS BiOp Action IV.1 and D-1641}

Compiled by Mike Urkov \email{mike.urkov@gmail.com}
}
\usage{
delta_cross_channel_closed
}
\description{
The number of days and proportion of days the Delta Cross Channel
gates are typically closed for each month
}
\details{
By rule, 45 days between November-January, based on real time monitoring.
For modeling purposes, the days closed where divided between December and January.

Note: Some real-time changes possible based on:
\itemize{
 \item fish monitoring
 \item interior delta salinity
 \item flood operations
}

In May, typically open for Memorial Day.
}
\keyword{datasets}
