% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_proportion_diverted}
\alias{delta_proportion_diverted}
\title{Delta Proportion Diverted}
\format{
A 3 dimensional array: 12 by 21 by 2 [months, years, deltas]
\describe{
\item{[ , , 1] North Delta}{represented with CALSIM II nodes (D403A + D403B + D403C + D403D + D404) / (C400 + C157)
north delta diversions divided by north delta inflow in cubic feet per second}

\item{[ , , 2] South Delta}{represented with CALSIM II nodes
(D418 + D419 + D412 + D410 + D413 + D409B + D416 + D408_OR + D408_VC) / (C401B + C504 + C508 + C644)
south delta diversions divided by south delta inflow in cubic feet per second}
}
}
\usage{
delta_proportion_diverted
}
\description{
The proportion of delta inflow diverted from 1980-2000.
}
\details{
The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River
below Vernalis.

Calculated using \code{\link{delta_flows}}
}
\keyword{datasets}
