% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gates_overtopped}
\alias{gates_overtopped}
\title{If gates are overtopped at teh Sutter and Yolo Bypasses}
\format{
a 12 by 21 by 2 array (month, year, bypass):

[ , , 1] = Sutter Bypass represented with CALSIM II nodes (D117 + D124 + D125 + D126)/C116

[ , , 2] = Yolo Bypass represented with CALSIM II nodes D160/C134
}
\source{
\itemize{
  \item \strong{Data Wrangling:} Erin Cain  \email{ecain@flowwest.com}
}
}
\usage{
gates_overtopped
}
\description{
A monthly TRUE or FALSE value to describe if the gates are overtopped to the bypasses
(years 1980-2000) for use with the CVPIA SIT Salmon Population Model to
apportion fish onto the bypasses.
}
\keyword{datasets}
