% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hatchery_flows}
\alias{hatchery_flows}
\alias{hatchery_apr_may_flows}
\alias{hatchery_oct_nov_flows}
\title{Hatchery Flows}
\usage{
hatchery_apr_may_flows

hatchery_oct_nov_flows
}
\description{
Flows in cfs for watersheds downstream of a hatchery.
}
\details{
Flows are derived from the \code{DSMflows::flows_cfs} dataset. Flows by year and watershed using the \code{mean}. These dataset are
used by the stray rate model in the DSM.
}
\keyword{datasets}
