% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean_flow}
\alias{mean_flow}
\title{Monthly Mean Flow (cms)}
\format{
a named list of two components, each containing a 3 dimensional array \link{31 watersheds, 12 months, 21 years}
}
\source{
Calculated using \code{\link{flows_cfs}}
}
\usage{
mean_flow
}
\description{
A named list containing the mean flow in cubic meters per second for each watershed every
month of every year in the simulation (1980-2000). The components of the list contain output from
three different Calsim runs:
\itemize{
\item biop_2008_2009: output from the 2008-2009 Calsim II BiOp run.
\item biop_itp_2018_2019: output from the 2018-2019 Calsim II BiOp run.
\item calsim_run_of_river: output from the Calsim Run of River run.
}
}
\keyword{datasets}
