% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportion_diverted}
\alias{proportion_diverted}
\title{Proportion of Flow Diverted}
\format{
a 31 by 12 by 21 array [watershed, month, year (1980-2000)]
}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{Node Selection:} Mark Tompkins \email{mtompkins@flowwest.com} and Mike Urkov \email{mike.urkov@gmail.com}
  \item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
proportion_diverted
}
\description{
A dataset containing the proportion of flow diverted within watersheds for
use with the CVPIA SIT Salmon Population Model.
}
\details{
The proportion diverted was calculated using 'FLOW-CHANNEL' and 'FLOW-DELIVERY' nodes from CALSIM II.
The nodes and calculation for each watershed are outlined below

The diversions of Antelope Creek, Deer Creek and Mill Creek are represented by one diversion node.
Diversions for these creeks were estimated in proportion to their flow. Elder Creek and Thomes Creek
are also represented with a single node and their diversions were estimated using the same method.

\describe{
  \item{Upper Sacramento River}{D104 / C104}
  \item{Antelope Creek}{(C11307 / (C11307 + C11308 + C11309) * D11305) / C11307}
  \item{Battle Creek}{no modeled diversions}
  \item{Bear Creek}{no modeled diversions}
  \item{Big Chico Creek}{no modeled diversions}
  \item{Butte Creek}{(C217B + D217) / (C217B + D217 + C217A)}
  \item{Clear Creek}{no modeled diversions}
  \item{Cottonwood Creek}{no modeled diversions}
  \item{Cow Creek}{no modeled diversions}
  \item{Deer Creek}{(C11309 / (C11307 + C11308 + C11309) * D11305) / C11309}
  \item{Elder Creek}{(C11303 / (C11303 + C11304) * D11301) / C11303}
  \item{Mill Creek}{(C11308 / (C11307 + C11308 + C11309) * D11305) / C11308}
  \item{Paynes Creek}{no modeled diversions}
  \item{Stony Creek}{D17301 / C42}
  \item{Thomes Creek}{(C11304 / (C11303 + C11304) * D11301) / C11304}
  \item{Upper-mid Sacramento River}{(D109 + D112 + D113A + D113B + D114 + D118 + D122A + D122B + D123 + D124A + D128_WTS + D128) / C110}
  \item{Sutter Bypass}{no modeled diversions}
  \item{Bear River}{D285 / (C285 + D285)}
  \item{Feather River}{(D201 + D202 + D7A + D7B) / C6}
  \item{Yuba River}{D230 / (C230 + D230)}
  \item{Lower-mid Sacramento River}{(D129A + D134 + D162 + D165) / C128}
  \item{Yolo Bypass}{no modeled diversions}
  \item{American River}{D302 / C9}
  \item{Lower Sacramento River}{(D167 + D168 + D168A_WTS) / C166}
  \item{Calaveras River}{(D506A + D506B + D506C + D507) / C92}
  \item{Cosumnes River}{no modeled diversions}
  \item{Mokelumne River}{(D503A + D503B + D503C + D502A + D502B) / C91*}
  \item{Merced River}{(D562 + D566) / C561}
  \item{Stanislaus River}{D528 / C520}
  \item{Tuolumne River}{D545 / C540}
  \item{San Joaquin River}{(D637 + D630B + D630A + D620B) / (D637 + D630B + D630A + D620B + C637)}
}


\emph{*Mokelumne River flow and diversions are from a separate model provided by EBMUD.}


\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}
\keyword{datasets}
