% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportion_pulse_flows}
\alias{proportion_pulse_flows}
\title{Proportion Pulse Flow}
\format{
a named list where each component is a 31 by 12 matrix (watersheds by months)
}
\usage{
proportion_pulse_flows
}
\description{
A named list with estimated monthly proportion of flow that is a pulse. Each component
of the list contains output from a different Calsim run:
\itemize{
\item biop_2008_2009: output from the 2008-2009 Calsim II BiOp run.
\item biop_itp_2018_2019: output from the 2018-2019 Calsim II BiOp run.
\item calsim_run_of_river: output from the Calsim Run of River run.
}
}
\details{
proportion_pulse_flows = sd(flow)/median(flow)/100

The value is divided by 100 to adjust the scaling to more closely resemble measured data.

Calculated using \code{\link{flows_cfs}}
}
\keyword{datasets}
