% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swp_exports}
\alias{swp_exports}
\title{SWP Exports}
\format{
A matrix with months (1-12) as rows and years (1980-1999) as columns
}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
swp_exports
}
\description{
Total exports for SWP in cms. Value is obtained using CALSIM variable DEL_SWP_EXP.
}
\details{
The flows here are obtained from a CALSIM II run designed for the structured
decision making in the delta.
\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}
\keyword{datasets}
