% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vernalis_flow}
\alias{vernalis_flow}
\title{Flow at Vernalis}
\format{
A named list where each element contains a matrix with months (1-12) as rows and years (1980-1999) as columns
}
\source{
\itemize{
\item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
\item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
vernalis_flow
}
\description{
flow in cms at Vernalis (CALSIM node C639) from two different Calsim II BiOp runs. Data
from each Calsim II BiOp run are stored as elements in a named list, named after the Calsim II BiOp run
they came from:
\itemize{
\item biop_2008_2009: output from the 2008-2009 Calsim II BiOp run.
\item biop_itp_2018_2019: output from the 2018-2019 Calsim II BiOp run.
}
}
\details{
\code{vernalis_flow} was calculated using a 'FLOW-CHANNEL' node from CALSIM II.

\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}
\keyword{datasets}
