% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bypass_flows}
\alias{bypass_flows}
\title{Flow through the Sutter and Yolo Bypasses}
\format{
dataframe with 972 rows and 7 variables:
\describe{
\item{date}{CALSIM II date}
\item{sutter1}{D117}
\item{sutter2}{C135}
\item{sutter3}{C136A}
\item{sutter4}{C137}
\item{yolo1}{D160}
\item{yolo2}{C157}
}
}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{Node Selection:} Mark Tompkins \email{mtompkins@flowwest.com}
  \item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
bypass_flows
}
\description{
A dataset containing the flow within the bypasses for
estimating available rearing habitat within the bypasses.
}
\details{
The flow in cubic feet per second through the bypasses are represented using
'FLOW-CHANNEL' and 'FLOW-DELIVERY' nodes from CALSIM II.
The nodes for each watershed are outlined above.
}
\section{Model Usage}{

The habitat within the bypasses are split at the major weirs.


\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}

\keyword{datasets}
