% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_proportion_diverted}
\alias{delta_proportion_diverted}
\title{Delta Proportion Diverted}
\format{
A 3 dimensional array: 12 by 21 by 2 [months, years, deltas]

[ , , 1] North Delta

[ , , 2] South Delta
}
\usage{
delta_proportion_diverted
}
\description{
The proportion of delta inflow diverted from 1980-2000.
}
\details{
The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River
below Vernalis.

Calculated using \code{\link{delta_flows}}
}
\keyword{datasets}
