% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_total_diverted}
\alias{delta_total_diverted}
\title{Delta Total Diverted}
\format{
A named list where each element contains a 3 dimensional array: 12 by 21 by 2 (months, years, deltas)

\link{ , , 1} North Delta

\link{ , , 2} South Delta
}
\usage{
delta_total_diverted
}
\description{
The total diverted of delta inflow in cubic meters per second from 1980-2000 from
three different Calsim runs. These can be accessed as separate elements of a named list named after
the Calsim run they came from:
\itemize{
\item biop_2008_2009: output from the 2008-2009 Calsim II BiOp run.
\item biop_itp_2018_2019: output from the 2018-2019 Calsim II BiOp run.
\item calsim_run_of_river: output from the Calsim Run of River run.
}
}
\details{
The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River
below Vernalis.

Calculated using \code{\link{delta_flows}}
}
\keyword{datasets}
