% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean_flow}
\alias{mean_flow}
\title{Monthly Mean Flow (cms)}
\format{
a 3 dimensional array [31 watersheds, 12 months, 21 years]
}
\source{
Calculated using \code{\link{flows_cfs}}
}
\usage{
mean_flow
}
\description{
The mean flow in cubic meters per second for each watershed every
month of every year in the simulation (1980-2000).
}
\keyword{datasets}
