% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportion_flow_bypasses}
\alias{proportion_flow_bypasses}
\title{Proportion of Sacramento River Flow through the Sutter and Yolo Bypasses}
\format{
a 12 by 21 by 2 array (month, year, bypass):

[ , , 1] = Sutter Bypass represented with CALSIM II nodes (D117 + D124 + D125 + D126)/C116

[ , , 2] = Yolo Bypass represented with CALSIM II nodes D160/C134
}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{Node Selection:} Mark Tompkins \email{mtompkins@flowwest.com} and Mike Urkov \email{mike.urkov@gmail.com}
  \item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
proportion_flow_bypasses
}
\description{
The monthly proportion of Sacramento River flow within the bypasses
(years 1980-2000) for use with the CVPIA SIT Salmon Population Model to
apportion fish onto the bypasses.
}
\details{
The proportions of Sacramento River flowing through the bypasses are represented using
'FLOW-CHANNEL' and 'FLOW-DELIVERY' nodes from CALSIM II.

 Some of the calculated values for Yolo Bypass were greater than 1.
 These values were capped at 1 for model usage.

\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}
\keyword{datasets}
