% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportion_pulse_flows}
\alias{proportion_pulse_flows}
\title{Proportion Pulse Flow}
\format{
a 31 by 12 matrix (watersheds by months)
}
\usage{
proportion_pulse_flows
}
\description{
Estimated monthly proportion of flow that is a pulse
}
\details{
proportion_pulse_flows = sd(flow)/median(flow)/100

The value is divided by 100 to adjust the scaling to more closely resemble measured data.

Calculated using \code{\link{flows_cfs}}
}
\keyword{datasets}
