% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stockton_flow}
\alias{stockton_flow}
\title{Flow at Stockton}
\format{
A matrix with months (1-12) as rows and years (1980-1999) as columns
}
\usage{
stockton_flow
}
\description{
flow in cms at Stockton (C417A)
}
\details{
The flows here are obtained from a CALSIM run designed for the structured
decision making in the delta. The name of the scenario is: Full Obs No Notch Baseline.
}
\keyword{datasets}
